#19 - Moved errorMessage function to a utils folder
