+++ src/modules/room/room.errors.ts
@@ -1,13 +1,3 @@
-export async function errorMessage(e: unknown) {
-    let message = String(e);
-
-    if (e instanceof Error) {
-        message = e.message;
-    }
-
-    return message;
-}
-
 export function idNotFound(id: number) {
     return Error("Could not find room with id: " + id);
 }
