+++ src/utils/string.ts
@@ -1,0 +1,9 @@
+export async function errorMessage(e: unknown) {
+    let message = String(e);
+
+    if (e instanceof Error) {
+        message = e.message;
+    }
+
+    return message;
+}
